requirements.txt
pip>=23.3
setuptools>=70.1.1
Django>=4.2.16,<4.3
djangorestframework>=3.15.2,<3.16
psycopg2-binary>=2.9.9,<3.0
drf-spectacular>=0.26.0,<0.27.0
Pillow>=10.3.0,<11.0.0
uwsgi>=2.0.20,<2.1.0
django-cors-headers>=4.3.1,<4.4.0
python-dotenv>=1.0.0,<1.1.0
drf-yasg==1.21.7
dj-database-url>=2.1.0,<2.2.0
whitenoise>=6.5.0,<7.0.0
django-ratelimit>=4.0.0,<4.1.0
PyJWT>=2.0.0
structlog>=23.1.0
djangorestframework-simplejwt[blacklist]>=5.3.0
pandas>=2.0.0,<3.0.0
celery>=5.3.0,<5.4.0
redis>=5.0.0,<6.0.0
django-celery-beat>=2.5.0,<2.6.0
django-redis>=5.0.0,<6.0.0
gunicorn>=21.0.0
sslcommerz-lib